Relocate header price into shared components
